+++ pennyway-client-iOS/pennyway-client-iOS/Api/Error/ErrorCode/ErrorCodeHandler/InternalServerError.swift
@@ -5,7 +5,8 @@ func internalServerError(_ code: String, message: String) -> StatusSpecificError
     }
 
     let defaultMessage = "Unexpected Error"
-    let fieldErrors = ErrorResponseData(field: internalServerError.rawValue)
+//    let fieldErrors = ErrorResponseData(field: internalServerError.rawValue)
+    let fieldErrors: ErrorResponseData? = nil
 
     return StatusSpecificError(domainError: .internalServerError, code: code, message: message.isEmpty ? defaultMessage : message, fieldErrors: fieldErrors)
 }
