+++ pennyway-client-iOS/pennyway-client-iOS/Api/Error/StatusCode/StatusSpecificError.swift
@@ -3,5 +3,5 @@ struct StatusSpecificError: Error {
     let domainError: StatusError
     let code: String
     let message: String
-    let fieldErrors: ErrorResponseData!
+    let fieldErrors: ErrorResponseData?
 }
