+++ pennyway-client-iOS/pennyway-client-iOS/View/CustomView/CustomToggleStyle.swift
@@ -18,7 +18,7 @@ struct CustomToggleStyle: ToggleStyle {
                         .shadow(radius: 1, x: 0, y: 1)
                         .padding(1.5)
                         .offset(x: configuration.isOn ? 10 : -10))
-                .animation(Animation.easeInOut(duration: 0.2))
+                .animation(Animation.easeInOut(duration: 0.3))
                 .onTapGesture { configuration.isOn.toggle() }
         }
     }
